+++ src/java/admin/controllers/DeactivateStudent.java
@@ -10,13 +10,11 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import org.hibernate.Session;
 import utility.Controller;
-import utility.Utils;
 
 /**
  *
@@ -32,7 +30,7 @@ public class DeactivateStudent extends Controller {
     @Override
     public void process(HttpServletRequest req, HttpServletResponse resp, Session session, HttpSession httpSession, PrintWriter out) throws Exception {
 
-        int studentId = Integer.parseInt(req.getParameter("studentId"));
+        long studentId = Long.parseLong(req.getParameter("studentId"));
         Student student = (Student) session.get(Student.class, studentId);
         student.setVerified(false);
