+++ src/java/admin/controllers/DetailClassRoom.java
@@ -10,13 +10,11 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import org.hibernate.Session;
 import utility.Controller;
-import utility.Utils;
 
 /**
  *
@@ -32,13 +30,14 @@ public class DetailClassRoom extends Controller {
     @Override
     public void process(HttpServletRequest req, HttpServletResponse resp, Session session, HttpSession httpSession, PrintWriter out) throws Exception {
 
-        int classRoomId = Integer.parseInt(req.getParameter("classroomId"));
+        long classRoomId = Long.parseLong(req.getParameter("classroomId"));
 
         ClassRoom classRoom = (ClassRoom) session.get(ClassRoom.class, classRoomId);
         
         req.setAttribute("classroom", classRoom);
         
-        req.getRequestDispatcher("/WEB-INF/admin/detailclassroom.jsp").include(req, resp);
+        req.getRequestDispatcher("/WEB-INF/admin/detailclassroom.jsp")
+                .include(req, resp);
 
     }
