+++ src/java/admin/postback/AddCourse.java
@@ -25,7 +25,7 @@ public class AddCourse extends PostBackController {
     @Override
     public void process(HttpServletRequest req, HttpServletResponse resp, Session session, HttpSession httpSession, PrintWriter out) throws Exception {
 
-        int departmentId = Integer.parseInt(req.getParameter("departmentId"));
+        long departmentId = Long.parseLong(req.getParameter("departmentId"));
         String name = req.getParameter("coursename");
 
         Department department = (Department) session.get(Department.class, departmentId);
